# Conflicts:
#	Kojiro/src/main/resources/templates/weakness.html

+++ Kojiro/src/main/java/com/example/Kojiro/controller/QManagementController.java
@@ -3,12 +3,11 @@ package com.example.Kojiro.controller;
 //QManagementController = QuestionManagementController(問題管理) 長くてすみません
 
 import com.example.Kojiro.dao.GenresDao;
-import com.example.Kojiro.entity.Questions2points;
-import com.example.Kojiro.entity.TestQuestion;
-import com.example.Kojiro.entity.TestQuestionP2;
-import com.example.Kojiro.entity.question;
+import com.example.Kojiro.entity.*;
 import com.example.Kojiro.form.QuizManagement;
+import com.example.Kojiro.form.QuizManagementStrAns;
 import com.example.Kojiro.service.QManagementService;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -16,10 +15,13 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardCopyOption;
 import java.util.Base64;
 
 @Controller
@@ -34,11 +36,14 @@ public class QManagementController {
     private GenresDao genresDao;
     @Autowired
     private HttpSession session;
+    @Autowired
+    private HttpServletRequest request;
+
     @GetMapping("/quiz-management")
     public String menu(@RequestParam(name="keyword" ,defaultValue = "") String keyword, Model model) {
-//        if (session.getAttribute("user")==null){//ユーザーのセッション判定
-//            return "redirect:/login-test";
-//        }
+        if(request.getSession(false)==null) return "redirect:/index";
+        var user = (Users)session.getAttribute("users");
+        if(user.role() != 1) return "redirect:/menu";
         if (keyword.isEmpty()){//検索欄にキーワードなしは全部出す
 //            model.addAttribute("management", qManagementService.findAll());
             model.addAttribute("management", qManagementService.findAll2());
@@ -72,6 +77,9 @@ public class QManagementController {
     @GetMapping("quiz-detail/{id}/{gName}")//id指定で詳細のページを出す
     public String detail(@PathVariable("id") int id,
                          @PathVariable("gName") String gName, Model model) {
+        if(request.getSession(false)==null) return "redirect:/index";
+        var user = (Users)session.getAttribute("users");
+        if(user.role() != 1) return "redirect:/menu";
         System.out.println(id);
         if(gName.equals("危険予測ディスカッション")){
             var result = qManagementService.findById2(id);
@@ -86,19 +94,6 @@ public class QManagementController {
         return "quiz-detail";
     }
 
-    public void showImage(String fName, Model model){
-        if(fName!=null && !fName.equals("")) {
-            File img = new File("./Kojiro/src/main/resources/static/images/" + fName);
-            try {
-                byte[] byteImg = Files.readAllBytes(img.toPath());
-                String base64Data = Base64.getEncoder().encodeToString(byteImg);
-                model.addAttribute("base64Data", "data:img/png;base64," + base64Data);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-
 //    @GetMapping("quiz-update/{id}")//id指定で更新のページを出す
 //    public String update1(@PathVariable("id") int id, Model model) {
 //        System.out.println(id);
@@ -109,6 +104,7 @@ public class QManagementController {
     @GetMapping("quiz-update/{id}/{genre}")//id指定で更新のページを出す
     public String update1(@PathVariable("id") int id,
                           @PathVariable("genre") String gName, Model model) {
+//        if(request.getSession(false)==null) return "redirect:/index";
         System.out.println(id);
         var genres = genresDao.findAll();
         if(gName.equals("危険予測ディスカッション")){
@@ -155,16 +151,21 @@ public class QManagementController {
 //        }
 //    }
     @PostMapping("quiz-update/{id}/{genre}")
-    public String update2(@PathVariable("id") int id, @PathVariable("genre") String gName, @Validated @ModelAttribute("q_management") QuizManagement update, BindingResult bindingResult, Model model){
+    public String update2(@PathVariable("id") int id, @PathVariable("genre") String gName, @Validated @ModelAttribute("q_management") QuizManagementStrAns update, BindingResult bindingResult, Model model){
+        if(request.getSession(false)==null) return "redirect:/index";
+        var user = (Users)session.getAttribute("users");
+        if(user.role() != 1) return "redirect:/menu";
         if(bindingResult.hasErrors()) {//バリデーションチェック
             return "quiz-update";
         }else {
+            // サーバーに画像を追加
+            insertImgFiles(update.getFile());
             if(gName.equals("危険予測ディスカッション")){
                 System.out.println(update);
                 var ansText = String.valueOf(update.getAnswer());
                 if(ansText.length() == 1) ansText="00"+ansText;
                 else if(ansText.length() == 2) ansText="0"+ansText;
-                var conProduct1 = new TestQuestionP2(id, update.getGenre(), update.getSentence(), ansText, update.getExplain(), update.getFile(), update.getScore());
+                var conProduct1 = new TestQuestionP2(id, update.getGenre(), update.getSentence(), ansText, update.getExplain(), update.getFile().getOriginalFilename(), update.getScore());
                 try {//正常に更新したとき
                     var result2 = qManagementService.update(conProduct1);
                     System.out.println(conProduct1);
@@ -177,9 +178,9 @@ public class QManagementController {
                 }
             } else {
                 System.out.println(update);
-                var conProduct1 = new TestQuestion(id, update.getGenre(), update.getSentence(), update.getAnswer(), update.getExplain(), update.getFile(), update.getScore());
+                var conProduct1 = new TestQuestion(id, update.getGenre(), update.getSentence(), update.getAnswer(), update.getExplain(), update.getFile().getOriginalFilename(), update.getScore());
                 try {//正常に更新したとき
-                    var result2 = qManagementService.update(conProduct1);
+                    qManagementService.update(conProduct1);
                     System.out.println(conProduct1);
                     successIndex = 2;
                     return "redirect:/quiz-management";
@@ -199,6 +200,9 @@ public class QManagementController {
 //    }
     @GetMapping("/quiz-delete/{id}/{genre}")//問題の削除(idを指定)
     public String delete1(@PathVariable("id") int id, @PathVariable("genre") String gName){
+        if(request.getSession(false)==null) return "redirect:/index";
+        var user = (Users)session.getAttribute("users");
+        if(user.role() != 1) return "redirect:/menu";
         if(gName.equals("危険予測ディスカッション")){
             var result = qManagementService.delete2(id);
             successIndex=3;
@@ -234,12 +238,17 @@ public class QManagementController {
 //        }
 //    }
     @PostMapping("/quiz-add")
-    public String product1(@Validated @ModelAttribute("add") QuizManagement add, BindingResult bindingResult) {
+    public String product1(@Validated @ModelAttribute("add") QuizManagementStrAns add, BindingResult bindingResult) {
+        if(request.getSession(false)==null) return "redirect:/index";
+        var user = (Users)session.getAttribute("users");
+        if(user.role() != 1) return "redirect:/menu";
         System.out.println(add);
         if (bindingResult.hasErrors()) {
             System.out.println("バリデーション");
             return "quiz-add";
         } else {
+            // 画像追加
+            insertImgFiles(add.getFile());
             try {
                 System.out.println("バリデーションなし");
 
@@ -248,10 +257,10 @@ public class QManagementController {
                     if(ansText.length() == 1) ansText="00"+ansText;
                     else if(ansText.length() == 2) ansText="0"+ansText;
                     System.out.println("アンサー：" + ansText);
-                    var result1 = qManagementService.insert(new Questions2points(-1, Integer.valueOf(add.getGenre()), add.getSentence(), ansText, add.getExplain(), add.getFile(), add.getScore()));
+                    qManagementService.insert(new Questions2points(-1, Integer.valueOf(add.getGenre()), add.getSentence(), ansText, add.getExplain(), add.getFile().getOriginalFilename(), add.getScore()));
                 } else {
-                    var conProduct = new TestQuestion(0, add.getGenre(), add.getSentence(), add.getAnswer(), add.getExplain(), add.getFile(), add.getScore());
-                    System.out.println(conProduct);
+                    var conProduct = new TestQuestion(0, add.getGenre(), add.getSentence(), add.getAnswer(), add.getExplain(), add.getFile().getOriginalFilename(), add.getScore());
+                    qManagementService.insert(conProduct);
                 }
                 successIndex = 1;
                 return "redirect:/quiz-management";
@@ -261,4 +270,34 @@ public class QManagementController {
             }
         }
     }
+
+
+    /* 画像関連メソッド */
+    // 画像表示
+    public void showImage(String fName, Model model){
+        if(fName!=null && !fName.equals("")) {
+            File img = new File("./Kojiro/src/main/resources/static/images/" + fName);
+            try {
+                byte[] byteImg = Files.readAllBytes(img.toPath());
+                String base64Data = Base64.getEncoder().encodeToString(byteImg);
+                model.addAttribute("base64Data", "data:img/png;base64," + base64Data);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+    }
+    // サーバーに画像を保存
+    public void insertImgFiles(MultipartFile file) {
+        final String UPLOAD_DIR = "./Kojiro/src/main/resources/static/images/";
+        try {
+            if (!file.getOriginalFilename().equals("")) {
+                String filePath = UPLOAD_DIR + File.separator + file.getOriginalFilename();
+                Path destination = new File(filePath).toPath();
+                Files.copy(file.getInputStream(), destination, StandardCopyOption.REPLACE_EXISTING);
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
 }
